requirements.txt
disnake==2.4.0
python-dotenv==0.20.0